Waiting for downstream job blocks jenkins
In case that there are more branches detected than executors
on nodes are available, this will deadlock jenkins completely,
as it tries to build another job while the current job is
still running and allocates an executor. If there are no executors
available, this will deadlock as the current job waits for the
scheduled downstream job forever.